rework architecture #2
